refactor: move parquet CDC writer-properties helper into write operation
writer_properties_from_format_options (and its PARQUET_CDC_* keys) is
write-path/format-specific and had only one caller, so it belongs next
to WriteBuilder rather than in the general table::config module. Also
drops a redundant EagerSnapshot::snapshot().metadata() hop in favor of
the direct metadata() accessor.
